feat(agent)!: expose usage through finish context
BREAKING CHANGE: @vite-hub/agent no longer exports usageTelemetry(), getUsageTelemetry(), usageTelemetry.from(), or the usage telemetry pricing/summary helper API. Agent usage is resolved through structured event.usage and context.usage data instead.
