+++ docs/content/blog/3.build-ai-chatbot.md
@@ -325,7 +325,7 @@ attach more Capabilities in the same list:
 - `kv()`, `blob()`, and `db()` expose scoped storage.
 - `llmGate()` rejects unsafe or unwanted requests.
 - `llmRoute()` chooses the right path for a request.
-- `usageTelemetry()` records model usage and cost.
+- Agent finish hooks can read structured usage from `event.usage`.
 
 The API should stay ordinary on purpose. Read the Agent Definition, see the
 model, see the Capabilities, and see the Workspace Sources the Agent can use.
