+++ docs/content/docs/agents/evals.md
@@ -96,6 +96,7 @@ Use `--watch` while editing prompts or scenarios. Use `--threshold` and `--outpu
 
 Good evals score source-grounded answers, refusal behavior, expected tool use, no source leakage, and regressions in usage or latency when telemetry is attached.
 When the behavior belongs to a Capability, assert its finish extension instead of duplicating host-specific hooks.
+Use `observability.usage` or finish-hook `event.usage` for structured Agent Usage Records.
 
 ```ts [server/agents/support.eval.ts]
 import { defineEval, hasCapabilityExtension } from '@vite-hub/agent/eval'
@@ -107,7 +108,7 @@ export default defineEval({
     await t.send('What changed in the order forecast?')
     t.hasCapabilityExtension('observability')
     t.expect(hasCapabilityExtension('observability', 'status'))
-    t.expect(hasCapabilityExtension('usage-telemetry'))
+    t.expect(hasCapabilityExtension('observability', 'usage'))
   },
 })
 ```
