+++ docs/content/docs/capabilities/observability.md
@@ -8,7 +8,7 @@ icon: i-lucide-radar
 ---
 
 `observability()` gives an Agent Definition one place to attach runtime telemetry.
-It can instrument model execution, emit lifecycle events, provide a finish extension with invocation status, duration, and result kind, and enable usage telemetry for the same invocation.
+It can instrument model execution, emit lifecycle events, and provide a finish extension with invocation status, duration, result kind, and structured usage when the Agent Driver reports it.
 
 ## Installation
 
@@ -37,14 +37,8 @@ When `onEvent` is configured, it also emits a `finish` or `error` event after th
 It provides an `observability` finish extension with `{ status, durationMs, resultKind, usage }` for completed invocations and `{ status, durationMs, usage }` for failed invocations.
 Agent Evals and the Agent test runner capture this finish extension automatically.
 
-`observability()` enables `usageTelemetry()` by default.
-When enabled usage telemetry records usage from the Agent Driver or custom result, `observability.usage` points at the same Agent Usage Record exposed through the `usage-telemetry` finish extension and final `result.usageRecord`.
-Use `observability({ usageTelemetry: false })` to opt out.
-With that opt-out, a custom `result.usageRecord` can still stay on the result, but `observability.usage` is not populated.
-
-If an Agent also lists `usageTelemetry(...)` in `capabilities`, that explicit Capability keeps its top-level order and wins for the Agent Usage Record itself.
-Place `usageTelemetry(...)` before `observability()` when `observability.usage` should reference the explicit record.
-`observability({ usageTelemetry: false })` still suppresses the `observability.usage` alias.
+Usage is exposed as a JSON-compatible Agent Usage Record. Finish hooks can read it from `event.usage`, and Channel Delivery finish effects can read it from `context.usage`.
+Applications own any formatting into text, chat messages, web UI, markdown, notes, billing records, or review comments.
 
 ## Eval assertions
 
@@ -74,7 +68,7 @@ For custom checks, read `observation.extensions.get('observability')` or `t.capa
 | --- | --- |
 | Model-backed | Supports model instrumentation, lifecycle events, finish extensions, and usage records when the model result reports usage. |
 | Harness-backed | Supports lifecycle events, finish extensions, and usage records when the harness reports usage; instrumentation applies only to model-backed execution. |
-| Custom-run-backed | Supports lifecycle events, finish extensions, and usage records when usage telemetry is enabled and the custom result reports usage. |
+| Custom-run-backed | Supports lifecycle events, finish extensions, and usage records when the custom result reports usage. |
 
 ## Options
 
@@ -82,7 +76,6 @@ For custom checks, read `observation.extensions.get('observability')` or `t.capa
 | --- | --- | --- | --- |
 | `instrumentation` | `AgentModelExecutionInstrumentation` | none | Model and call-settings instrumentation for model-backed drivers. |
 | `onEvent` | `(event) => void` | none | Lifecycle event sink for start, finish, and error events. |
-| `usageTelemetry` | `boolean \| UsageTelemetryOptions` | `true` | Enables usage telemetry by default, accepts inline usage telemetry options, or opts out when set to `false`. |
 
 ## Reference
