+++ docs/content/docs/capabilities/official-capabilities.md
@@ -37,7 +37,6 @@ import {
   skills,
   subagents,
   transcribe,
-  usageTelemetry,
   webSearch,
   workspaceShell,
 } from '@vite-hub/agent/capabilities'
@@ -73,7 +72,6 @@ import {
 | Rate limit | [`rateLimit()`](/docs/capabilities/rate-limit) | A trusted invocation budget should be checked or consumed before the Agent runs. |
 | Chat title | [`chatTitle()`](/docs/capabilities/chat-title) | Chat streams and finish extensions should include a generated conversation title. |
 | Chat summary | [`chatSummary()`](/docs/capabilities/chat-summary) | A summary command should replace explicit input with a conversation summary. |
-| Usage telemetry | [`usageTelemetry()`](/docs/capabilities/usage-telemetry) | Agent Usage Records should be normalized, emitted, or attached to output. |
 | Observability | [`observability()`](/docs/capabilities/observability) | Lifecycle events, model instrumentation, and finish metadata should be attached to Agent Invocations. |
 
 ## Read capability pages first
