+++ packages/agent/src/index.ts
@@ -9,7 +9,7 @@ import {
 } from "./delivery-effects.ts"
 import { getMessageText } from "./messages.ts"
 import { resolveRuntimeContext } from "@vite-hub/runtime"
-import { isAsyncIterable, streamAgentOutputToEvents, toAgentRunResult, toAgentStreamEvent } from "./agent-output.ts"
+import { isAsyncIterable, resolveAgentUsageRecord, streamAgentOutputToEvents, toAgentRunResult, toAgentStreamEvent } from "./agent-output.ts"
 import { defineChatCapability, getChatCapabilityOptions } from "./chat-trigger.ts"
 import { resolveAgentChannelChatOptions } from "./internal/channels.ts"
 import {
@@ -1843,6 +1843,7 @@ function createFinishDeliveryEffectContext<
     run: context.run,
     status: createStatusDeliveryEffectIntent,
     ...(event.text !== undefined ? { text: event.text } : {}),
+    ...(event.usage !== undefined ? { usage: event.usage } : {}),
     workspace: context.workspace,
   }
 }
@@ -1911,7 +1912,9 @@ async function finishAgentInvocation<
   const failed = outcome.status === "error"
   const error = failed ? outcome.error : undefined
   const result = outcome.status === "success" ? outcome.result : undefined
-  const text = failed || result === undefined ? undefined : toAgentRunResult(result).text
+  const runResult = failed || result === undefined ? undefined : toAgentRunResult(result)
+  const text = runResult?.text
+  const usage = result === undefined ? undefined : await resolveAgentUsageRecord(result, context.run)
   try {
     await context.close()
     if (hasFinishWork(context)) {
@@ -1929,6 +1932,7 @@ async function finishAgentInvocation<
         ...(result !== undefined ? { result } : {}),
         runtime: context.runtimeContext,
         ...(text !== undefined ? { text } : {}),
+        ...(usage !== undefined ? { usage } : {}),
       } satisfies Omit<AgentFinishEvent<TRuntimeConfig, CALL_OPTIONS>, "extensions">
       context.context.set("agent.finishHook", Boolean(context.finishHook), { overwrite: true })
       const provisionalActiveDeliveryProviders = await prepareProvisionalTitleDeliverySupport(context, eventBase)
