+++ packages/agent/test/providers.test.ts
@@ -2011,7 +2011,7 @@ describe("server helpers", () => {
 
   it("handles Chat SDK webhooks through the chat capability", async () => {
     const { defineAgent } = await import("../src/index.ts")
-    const { access, staticModelPricing, usageTelemetry } = await import("../src/capabilities.ts")
+    const { access } = await import("../src/capabilities.ts")
     const { defineChatCapability } = await import("../src/chat-trigger.ts")
     const { createChannelWebhookRouteHandler } = await import("../src/server/internal.ts")
     const adapter = createTestChatAdapter()
@@ -2056,14 +2056,6 @@ describe("server helpers", () => {
           webhooks: {
             telegram: {},
           },
-        }),
-        usageTelemetry({
-          pricing: staticModelPricing({
-            "openai/gpt-test": {
-              input: "0.00000010",
-              output: "0.00000020",
-            },
-          }),
         }),
       ],
       invoker: {
@@ -4477,14 +4469,13 @@ describe("server helpers", () => {
     }
   })
 
-  it("lets agent finish hooks post usage telemetry for non-streaming model chat webhooks", async () => {
-    const { staticModelPricing, usageTelemetry } = await import("../src/capabilities.ts")
+  it("lets agent finish hooks post usage for non-streaming model chat webhooks", async () => {
     const { defineChatCapability } = await import("../src/chat-trigger.ts")
     const { createChannelWebhookRouteHandler } = await import("../src/server/internal.ts")
     const adapter = createTestChatAdapter()
     const finish = vi.fn(async (event) => {
       const chat = event.extensions.get("chat") as { provider?: string, sendMessage?: (message: { markdown: string }) => Promise<void> } | undefined
-      const usage = event.extensions.get("usage-telemetry") as { usage?: { totalTokens?: number } } | undefined
+      const usage = event.usage
       if (chat && usage) {
         await chat.sendMessage?.({
           markdown: `Custom usage: \`${usage.usage?.totalTokens}\` tokens via ${chat.provider}`,
@@ -4521,14 +4512,6 @@ describe("server helpers", () => {
             telegram: {},
           },
         }),
-        usageTelemetry({
-          pricing: staticModelPricing({
-            "openai/gpt-test": {
-              input: "0.00000010",
-              output: "0.00000020",
-            },
-          }),
-        }),
       ],
       hooks: {
         "agent:finish": finish,
@@ -4558,11 +4541,7 @@ describe("server helpers", () => {
     expect(adapter.postMessage).toHaveBeenNthCalledWith(1, "telegram:456", { markdown: "generated ok" })
     expect(adapter.postMessage).toHaveBeenNthCalledWith(2, "telegram:456", { markdown: "Custom usage: `15` tokens via telegram" })
     expect(finish).toHaveBeenCalledOnce()
-    expect(finish.mock.calls[0]![0].extensions.get("usage-telemetry")).toEqual(expect.objectContaining({
-      cost: expect.objectContaining({
-        amount: "0.0000018",
-        currency: "USD",
-      }),
+    expect(finish.mock.calls[0]![0].usage).toEqual(expect.objectContaining({
       latency: expect.objectContaining({
         durationMs: 900,
       }),
@@ -5047,7 +5026,6 @@ describe("server helpers", () => {
 
   it("flushes deferred non-streaming chat webhook work before returning", async () => {
     const { defineAgent } = await import("../src/index.ts")
-    const { usageTelemetry } = await import("../src/capabilities.ts")
     const { defineChatCapability } = await import("../src/chat-trigger.ts")
     const { createChannelWebhookRouteHandler } = await import("../src/server/internal.ts")
     const adapter = createTestChatAdapter({ deferMessageProcessing: true })
@@ -5073,7 +5051,7 @@ describe("server helpers", () => {
     })
     const finish = vi.fn(async (event) => {
       const chat = event.extensions.get("chat") as { sendMessage?: (message: { markdown: string }) => Promise<void> } | undefined
-      const usage = event.extensions.get("usage-telemetry")
+      const usage = event.usage
       if (chat && usage) {
         await chat.sendMessage?.({ markdown: "usage ok" })
       }
@@ -5089,7 +5067,6 @@ describe("server helpers", () => {
             telegram: {},
           },
         }),
-        usageTelemetry(),
       ],
       hooks: {
         "agent:finish": finish,
@@ -5132,15 +5109,15 @@ describe("server helpers", () => {
     expect(finish).toHaveBeenCalledOnce()
   })
 
-  it("lets agent finish hooks compose usage telemetry and chat follow-up messages", async () => {
+  it("lets agent finish hooks compose usage and chat follow-up messages", async () => {
     const { defineAgent } = await import("../src/index.ts")
-    const { access, staticModelPricing, usageTelemetry } = await import("../src/capabilities.ts")
+    const { access } = await import("../src/capabilities.ts")
     const { defineChatCapability } = await import("../src/chat-trigger.ts")
     const { createChannelWebhookRouteHandler } = await import("../src/server/internal.ts")
     const adapter = createTestChatAdapter()
     const finish = vi.fn(async (event) => {
       const chat = event.extensions.get("chat") as { provider?: string, sendMessage?: (message: { markdown: string }) => Promise<void> } | undefined
-      const usage = event.extensions.get("usage-telemetry") as { usage?: { totalTokens?: number } } | undefined
+      const usage = event.usage
       if (chat && usage) {
         await chat.sendMessage?.({
           markdown: `Custom usage: \`${usage.usage?.totalTokens}\` tokens via ${chat.provider}`,
@@ -5161,14 +5138,6 @@ describe("server helpers", () => {
           webhooks: {
             telegram: {},
           },
-        }),
-        usageTelemetry({
-          pricing: staticModelPricing({
-            "openai/gpt-test": {
-              input: "0.00000010",
-              output: "0.00000020",
-            },
-          }),
         }),
       ],
       hooks: {
@@ -5207,7 +5176,7 @@ describe("server helpers", () => {
     expect(adapter.editMessage).toHaveBeenCalledWith("telegram:789", "sent-1", { markdown: "ok" })
     expect(adapter.postMessage).toHaveBeenNthCalledWith(2, "telegram:789", { markdown: "Custom usage: `15` tokens via telegram" })
     expect(finish).toHaveBeenCalledOnce()
-    expect(finish.mock.calls[0]![0].extensions.get("usage-telemetry")).toEqual(expect.objectContaining({
+    expect(finish.mock.calls[0]![0].usage).toEqual(expect.objectContaining({
       usage: {
         inputTokens: 10,
         outputTokens: 5,
