+++ packages/vite/src/agent/capabilities.ts
@@ -22,7 +22,6 @@ export {
   mcp,
   memory,
   memoryRateLimitStore,
-  normalizeAgentUsage,
   observability,
   openapi,
   pullRequestContext,
@@ -31,11 +30,8 @@ export {
   sandbox,
   schedule,
   skills,
-  staticModelPricing,
   subagents,
   transcribe,
-  usageTelemetry,
-  vercelAiGatewayPricing,
   webSearch,
   workspaceJsonlMemoryStore,
   workspaceShell,
